requirements.txt
chromadb
mistralai
pdfplumber
python-dotenv
spacy
pymupdf

